+++ poem-openapi/src/types/base64_type.rs
@@ -1,5 +1,9 @@
-use std::borrow::Cow;
+use std::{
+    borrow::Cow,
+    ops::{Deref, DerefMut},
+};
 
+use bytes::Bytes;
 use serde_json::Value;
 
 use crate::{
@@ -9,9 +13,23 @@ use crate::{
 
 /// Represents a binary data encoded with base64.
 #[derive(Debug, Clone, Eq, PartialEq, Hash)]
-pub struct Base64(pub Vec<u8>);
+pub struct Base64<T>(pub T);
 
-impl Type for Base64 {
+impl<T> Deref for Base64<T> {
+    type Target = T;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl<T> DerefMut for Base64<T> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
+impl<T: Send + Sync> Type for Base64<T> {
     const IS_REQUIRED: bool = true;
 
     type RawValueType = Self;
@@ -37,7 +55,7 @@ impl Type for Base64 {
     }
 }
 
-impl ParseFromJSON for Base64 {
+impl ParseFromJSON for Base64<Vec<u8>> {
     fn parse_from_json(value: Value) -> ParseResult<Self> {
         if let Value::String(value) = value {
             Ok(Self(base64::decode(value)?))
@@ -47,14 +65,30 @@ impl ParseFromJSON for Base64 {
     }
 }
 
-impl ParseFromParameter for Base64 {
+impl ParseFromJSON for Base64<Bytes> {
+    fn parse_from_json(value: Value) -> ParseResult<Self> {
+        if let Value::String(value) = value {
+            Ok(Self(base64::decode(value).map(Into::into)?))
+        } else {
+            Err(ParseError::expected_type(value))
+        }
+    }
+}
+
+impl ParseFromParameter for Base64<Vec<u8>> {
     fn parse_from_parameter(value: &str) -> ParseResult<Self> {
         Ok(Self(base64::decode(value)?))
     }
 }
 
-impl ToJSON for Base64 {
+impl ParseFromParameter for Base64<Bytes> {
+    fn parse_from_parameter(value: &str) -> ParseResult<Self> {
+        Ok(Self(base64::decode(value).map(Into::into)?))
+    }
+}
+
+impl<T: AsRef<[u8]> + Send + Sync> ToJSON for Base64<T> {
     fn to_json(&self) -> Value {
-        Value::String(base64::encode(&self.0))
+        Value::String(base64::encode(self.0.as_ref()))
     }
 }