+++ poem-openapi/src/types/binary.rs
@@ -1,5 +1,9 @@
-use std::borrow::Cow;
+use std::{
+    borrow::Cow,
+    ops::{Deref, DerefMut},
+};
 
+use bytes::Bytes;
 use poem::web::Field;
 
 use crate::{
@@ -9,9 +13,23 @@ use crate::{
 
 /// Represents a binary data.
 #[derive(Debug, Clone, Eq, PartialEq, Hash)]
-pub struct Binary(pub Vec<u8>);
+pub struct Binary<T>(pub T);
 
-impl Type for Binary {
+impl<T> Deref for Binary<T> {
+    type Target = T;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl<T> DerefMut for Binary<T> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
+impl<T: Send + Sync> Type for Binary<T> {
     const IS_REQUIRED: bool = true;
 
     type RawValueType = Self;
@@ -38,10 +56,26 @@ impl Type for Binary {
 }
 
 #[poem::async_trait]
-impl ParseFromMultipartField for Binary {
+impl ParseFromMultipartField for Binary<Vec<u8>> {
     async fn parse_from_multipart(field: Option<Field>) -> ParseResult<Self> {
         match field {
             Some(field) => Ok(Self(field.bytes().await.map_err(ParseError::custom)?)),
+            None => Err(ParseError::expected_input()),
+        }
+    }
+}
+
+#[poem::async_trait]
+impl ParseFromMultipartField for Binary<Bytes> {
+    async fn parse_from_multipart(field: Option<Field>) -> ParseResult<Self> {
+        match field {
+            Some(field) => Ok(Self(
+                field
+                    .bytes()
+                    .await
+                    .map(Into::into)
+                    .map_err(ParseError::custom)?,
+            )),
             None => Err(ParseError::expected_input()),
         }
     }