+++ poem-openapi/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- Add `Attachment` payload for download file.
+- Added `BinaryStream` to support streaming payload.
+
 # [1.0.46] 2021-12-10
 
 - Change the default renaming rule of enum items from `ScreamingSnake` to `Pascal`.